requirements.txt
fastapi
uvicorn[standard]
aiofiles
httpx
jinja2
aiogram==3.*
python-dotenv
